Refactoring the packages
