+++ dvd-store/src/main/java/com/quality/workshop/rest/VideoTapeResource.java
@@ -1,4 +1,4 @@
-package com.worldline.ep2.kata.rest;
+package com.quality.workshop.rest;
 
 import java.util.List;
 
@@ -6,8 +6,8 @@ import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 
-import com.worldline.ep2.kata.entity.Schauspieler;
-import com.worldline.ep2.kata.entity.VideoTape;
+import com.quality.workshop.entity.Schauspieler;
+import com.quality.workshop.entity.VideoTape;
 
 @Path("/video")
 public class VideoTapeResource {
